commit changes

+++ src/app/reservations/reservations.component.ts
@@ -16,63 +16,70 @@ export class ReservationsComponent implements OnInit {
   reservations: Reservation[];
   selectedReservation: Reservation;
 
-   
 
-  @Input() reservation:Reservation;
- 
-  constructor(private reservationService : ReservationService, private router: Router) { }
+  @Input() reservation: Reservation;
+
+  constructor(private reservationService: ReservationService, private router: Router) { }
+
 
   ngOnInit() {
-  this.reservationService.getReservationsByUser().subscribe(reservations=>{
-    
-      this.reservations=reservations;
+    this.reservationService.getReservationsByUser().subscribe(reservations => {
+      reservations.forEach(res => {
+        res.bookingStart = isNull(res.bookingStart) ? new Date() : (isDate(res.bookingStart) ? moment(Number(res.bookingStart)).toDate() : new Date());
+        res.bookingEnd = isNull(res.bookingEnd) ? new Date() : (isDate(res.bookingEnd) ? moment(Number(res.bookingEnd)).toDate() : new Date());
+      });
+      this.reservations = reservations;
     });
+
 
   }
 
-  deleteReservation(reservation:Reservation){
+  deleteReservation(reservation: Reservation) {
     console.log('Delete Reservation !');
-    this.reservationService.deleteReservation(reservation).subscribe(
-      status =>{
-        if(status){
+    if (confirm('Are you sure you want to delete this?')) {
+      // TODO:  Do something here if the answer is "Ok".
+      this.reservationService.deleteReservation(reservation).subscribe(
+        status => {
+          if (status) {
             this.reservations = this.reservations.filter(rs => rs.tableReservationId !== reservation.tableReservationId);
           }
         }
       );
-      }
-
-getAvailableTables(){
-    console.log('Get Available Tables to Book !');
-     this.reservationService.getAvailableTables().subscribe(reservations=>{
-
-      this.reservations=reservations;
-    });
-   
+    }
   }
 
+  getAvailableTables() {
+    console.log('Get Available Tables to Book !');
+    this.reservationService.getAvailableTables().subscribe(reservations => {
+      reservations.forEach(res => {
+        res.bookingStart = isNull(res.bookingStart) ? new Date() : (isDate(res.bookingStart) ? moment(Number(res.bookingStart)).toDate() : new Date());
+        res.bookingEnd = isNull(res.bookingEnd) ? new Date() : (isDate(res.bookingEnd) ? moment(Number(res.bookingEnd)).toDate() : new Date());
+      });
+      this.reservations = reservations;
+    });
 
+  }
 
-  addReservation(reservation:Reservation){
+  addReservation(reservation: Reservation) {
     console.log('Add  Reservation Service CAll  !');
     this.reservationService.addReservation(reservation)
-    .subscribe(
-      reservation=>this.reservations.push(reservation)
+      .subscribe(
+        reservation => this.reservations.push(reservation)
       );
-      this.selectedReservation=null;
-      this.router.navigate(['home']);
+    this.selectedReservation = null;
+    this.router.navigate(['/']);
   }
 
-  editReservation(reservation:Reservation){
+  editReservation(reservation: Reservation) {
     console.log('Edit Reservation ! @ Edit Page');
-    reservation.bookingStart = isNull(reservation.bookingStart) ? new Date() : (isDate(reservation.bookingStart) ? moment(Number(reservation.bookingStart)).toDate() : new Date() );
-     reservation.bookingEnd  = isNull(reservation.bookingEnd) ? new Date() : (isDate(reservation.bookingEnd) ? moment(Number(reservation.bookingEnd)).toDate() : new Date() );
-      
-    this.selectedReservation=reservation;
- 
+
+    reservation.bookingStart = isNull(reservation.bookingStart) ? new Date() : (isDate(reservation.bookingStart) ? moment(Number(reservation.bookingStart)).toDate() : new Date());
+    reservation.bookingEnd = isNull(reservation.bookingEnd) ? new Date() : (isDate(reservation.bookingEnd) ? moment(Number(reservation.bookingEnd)).toDate() : new Date());
+    this.selectedReservation = reservation;
   }
 
 
 
-    
-  
+
+
 }
